// src/client/types.gen.ts
export type ProjectCreate = {
    /**
     * Name
     */
    name: string;
    /**
     * Description
     */
    description?: string | null;
    /**
     * Deadline
     */
    deadline?: string | null;
    /**
     * Priority
     */
    priority?: string | null;
    /**
     * Knows Steps
     */
    knows_steps?: boolean | null;
    /**
     * Status
     */
    status?: string | null;
    /**
     * Color
     */
    color?: string | null;
    /**
     * Is Active
     */
    is_active?: boolean;
};